# server/api/swagger.yaml
basePath: /api/v1
definitions:
  model.DebugControlDispatch:
    properties:
      action:
        type: string
      agent_id:
        type: string
      payload:
        additionalProperties: {}
        type: object
    type: object
  model.DebugTaskDispatch:
    properties:
      agent_id:
        type: string
      command:
        type: string
      task_id:
        type: string
      timeout:
        type: integer
    type: object
  model.DeviceInfo:
    properties:
      arch:
        type: string
      hostname:
        type: string
      ip:
        type: string
      os:
        type: string
    type: object
  model.Envelope:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      request_id:
        type: string
    type: object
  model.HealthResp:
    properties:
      service:
        type: string
      status:
        type: string
      timestamp:
        type: integer
    type: object
  model.HeartbeatRequest:
    properties:
      agent_id:
        type: string
      metrics:
        $ref: '#/definitions/model.MetricsInfo'
      running_tasks:
        items:
          type: string
        type: array
      timestamp:
        type: integer
    type: object
  model.MetricsInfo:
    properties:
      cpu_percent:
        type: number
      disk_percent:
        type: number
      mem_percent:
        type: number
    type: object
  model.RegisterRequest:
    properties:
      agent_id:
        type: string
      agent_version:
        type: string
      capabilities:
        items:
          type: string
        type: array
      device:
        $ref: '#/definitions/model.DeviceInfo'
      device_code:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      tenant_id:
        type: string
    type: object
  model.ReportResult:
    properties:
      exit_code:
        type: integer
      stderr:
        type: string
      stdout:
        type: string
      truncated:
        type: boolean
    type: object
  model.TaskReportRequest:
    properties:
      agent_id:
        type: string
      event_id:
        type: string
      finished_at:
        type: integer
      result:
        $ref: '#/definitions/model.ReportResult'
      started_at:
        type: integer
      status:
        type: string
      task_id:
        type: string
    type: object
  model.TaskStatusRequest:
    properties:
      agent_id:
        type: string
      attempt:
        type: integer
      event_id:
        type: string
      status:
        type: string
      task_id:
        type: string
      timestamp:
        type: integer
    type: object
host: localhost:40001
info:
  contact: {}
  description: 远程 Agent 管理服务，提供注册、心跳、任务下发与上报等接口。
  title: Luoyi Remote Agent Server API
  version: "1.0"
paths:
  /api/v1/agent/heartbeat:
    post:
      consumes:
      - application/json
      parameters:
      - description: 心跳请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.HeartbeatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Envelope'
      security:
      - BearerAuth: []
      summary: Agent 心跳上报
      tags:
      - agent
  /api/v1/agent/poll:
    get:
      parameters:
      - description: Agent ID
        in: query
        name: agent_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Envelope'
      security:
      - BearerAuth: []
      summary: Agent 长轮询获取任务
      tags:
      - agent
  /api/v1/agent/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 注册请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Envelope'
      security:
      - AdminAuth: []
      summary: Agent 注册
      tags:
      - agent
  /api/v1/agent/task/report:
    post:
      consumes:
      - application/json
      parameters:
      - description: 任务报告请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.TaskReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Envelope'
      security:
      - BearerAuth: []
      summary: 上报任务执行结果
      tags:
      - agent
  /api/v1/agent/task/status:
    post:
      consumes:
      - application/json
      parameters:
      - description: 任务状态请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.TaskStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Envelope'
      security:
      - BearerAuth: []
      summary: 上报任务状态
      tags:
      - agent
  /api/v1/debug/dispatch/control:
    post:
      consumes:
      - application/json
      parameters:
      - description: 控制指令请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.DebugControlDispatch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Envelope'
      security:
      - AdminAuth: []
      summary: 调试：下发控制指令
      tags:
      - debug
  /api/v1/debug/dispatch/task:
    post:
      consumes:
      - application/json
      parameters:
      - description: 任务下发请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.DebugTaskDispatch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Envelope'
      security:
      - AdminAuth: []
      summary: 调试：下发任务
      tags:
      - debug
  /api/v1/debug/state:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Envelope'
      security:
      - AdminAuth: []
      summary: 调试：查看内存状态统计
      tags:
      - debug
  /healthz:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/model.HealthResp'
              type: object
      summary: 健康检查
      tags:
      - public
securityDefinitions:
  AdminAuth:
    in: header
    name: X-Register-Token
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"